feat: implement notification flow
